cambio a double
cambia a double

+++ src/main/java/com/homie/backend/sisInterno/dto/BusquedaDtoIn.java
@@ -11,7 +11,7 @@ public interface BusquedaDtoIn {
 
 	String getPeEstado();
 
-	Long getPeCantidadHoras();
+	Double getPeCantidadHoras();
 
 	String getPeTipo();
